refactor: move example codes out of document's source code
