+++ examples/models/googlenet.py
