+++ examples/models/squeezenet.py
